SampleSearch in it's own component
